Move RPC Class to it's own file
